+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   (InspectorCaracal)
 - Fix: `find #dbref` results didn't include the results of
   `.get_extra_display_name_info` (the #dbref display by default) (Griatch)
+- Fix: Add `DefaultAccount.get_extra_display_name_info` method for API
+  compliance with `DefaultObject` in commands. (Griatch)
 
 [pull3438]: https://github.com/evennia/evennia/pull/3446